Rename files so that the names are shorter and easier to read
