+++ ui/src/components/WMWorkerTypesTable/index.jsx
