+++ ui/src/views/WorkerManager/WMViewWorkerTypes/WMWorkerTypes.graphql
