+++ ui/src/views/WorkerManager/WMViewWorkers/WMWorkers.graphql
